Restructuring the module cnn
